+++ sktime_dl/deeplearning/tests/test_accuracy.py
@@ -11,7 +11,7 @@ import pytest
 from flaky import flaky
 from sktime.datasets import load_italy_power_demand
 
-from sktime_dl.deeplearning import CNNClassifier
+from sktime_dl.classification import CNNClassifier
 from sktime_dl.deeplearning import EncoderClassifier
 from sktime_dl.deeplearning import FCNClassifier
 from sktime_dl.deeplearning import InceptionTimeClassifier
