+++ sktime_dl/deeplearning/tests/test_is_fitted.py
@@ -4,7 +4,7 @@ from sktime.exceptions import NotFittedError
 from sktime.datasets import load_italy_power_demand
 from sktime.regression.base import BaseRegressor
 
-from sktime_dl.deeplearning import CNNClassifier
+from sktime_dl.classification import CNNClassifier
 from sktime_dl.utils.model_lists import SMALL_NB_EPOCHS
 from sktime_dl.utils.model_lists import construct_all_classifiers
 from sktime_dl.utils.model_lists import construct_all_regressors
